+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hackathon7",
+  "name": "yes",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -50,5 +50,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "main": "index.js",
+  "repository": "https://github.com/Hamidswed/hackathon7.git",
+  "author": "<mtgnatali@gmail.com>",
+  "license": "MIT"
 }
