+++ src/Components/allTasks/Alltasks.tsx
@@ -1,4 +1,6 @@
-import React, { useState } from 'react'
+import React, { FormEvent, useState } from 'react'
+import {Stack, TextField, Button} from '@mui/material/';
+
 import { SingleTaskInterface, tasksDefault } from '../../types/tasksType'
 import SingleTask from '../singleTask/SingleTask'
 
@@ -11,7 +13,14 @@ interface AllTasksInterface {
 
 const Alltasks = ({taskListArray,onDelete,onToggleDone, onToggleImportant}: AllTasksInterface ) => {
   const [task, setTask] = useState<SingleTaskInterface[] | []>(tasksDefault);
-  
+  const [textForNewTask, setTextForNewTask] = useState('');
+
+  const onSubmitTaskForm= (event : FormEvent<HTMLFormElement> ) => {
+    event.preventDefault();
+    addNewTask(textForNewTask);
+    setTextForNewTask('');
+}
+
   function getRandomArbitrary(min: number, max: number) {
     return Math.random() * (max - min) + min;
 }
@@ -60,6 +69,32 @@ const addNewTaskInState = (subtitle: string) => {
                 )}
             </ul>
         </section>
+        <Stack
+      component="form"
+      sx={{
+        width: '25ch',
+      }}
+      spacing={2}
+      noValidate
+      autoComplete="off"
+      onSubmit={onSubmitTaskForm}
+    >
+<TextField
+        hiddenLabel
+        id="filled-hidden-label-normal"
+        defaultValue="Normal"
+        variant="filled"
+        type="text"
+        placeholder="What do you want to do?"
+        value={textForNewTask}
+        onChange={(e) => {setTextForNewTask(e.target.value)}}
+      />
+      <Button variant="contained" color="success">
+      Add new task
+      </Button>
+       {/* <button type="submit" className="btn">Add new task</button> */}
+    </Stack>
+      
     </div>
   )
 }
