+++ src/Components/singleTask/SingleTask.tsx
@@ -1,4 +1,4 @@
-import React, { FormEvent, useState } from 'react'
+import React from 'react'
 import DeleteIcon from '../../images/icons/DeleteIcon'
 import DoneIcon from '../../images/icons/DoneIcon'
 import ImportantIcon from '../../images/icons/ImportantIcon'
@@ -16,13 +16,6 @@ interface SingleTaskInterface {
 }
 
 const SingleTask = ({name,onDelete,onToggleDone, onToggleImportant, addNewTask, important, done}: SingleTaskInterface)  => {
-  const [textForNewTask, setTextForNewTask] = useState('');
-
-  const onSubmitTaskForm= (event : FormEvent<HTMLFormElement> ) => {
-    event.preventDefault();
-    addNewTask(textForNewTask);
-    setTextForNewTask('');
-}
   
   return (
     <div className='task-container'>
@@ -40,16 +33,7 @@ const SingleTask = ({name,onDelete,onToggleDone, onToggleImportant, addNewTask, 
                 <button type="button" className="btn btn-outline-danger ms-1" onClick={onDelete}>
                     <DeleteIcon/>
                 </button>
-                <form action="" onSubmit={onSubmitTaskForm} className="task-form">
-                <input
-                    type="text"
-                    className="form-control"
-                    placeholder="What else do you want to do?"
-                    value={textForNewTask}
-                    onChange={(e) => {setTextForNewTask(e.target.value)}}
-                />
-                <button type="submit" className="btn">Add new task</button>
-                </form>
+                
             </div>
 
         </li>
