+++ src/Components/singleTask/singleTask.css
@@ -14,3 +14,8 @@
     border-radius:0 0 10px 10px;
 }
 
+.task-form {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+}
