+++ src/types/tasksType.ts
@@ -19,3 +19,4 @@ export interface SingleTaskInterface  {
     done: boolean,
 
 };
+
